+++ CHANGELOG.md
@@ -4,11 +4,17 @@
 
 ### Added
 
+- Add a Cigar format
+
 ### Changed
+
+- Replace derive_builder by typed_builder
 
 ### Deprecated
 
 ### Removed
+
+- All error related to Builder
 
 ### Fixed